Topic disambiguation "sub-pages" in HTML doc.
The purpose of this change is to handle situations where a
symbol has multiple meanings, resulting in different topic
areas of the document where we might want to navigate. For
instance, there is an equal function in the TXR Lisp library,
and an equal method that users can define.

* genman.txr (tochash, disamb, dist-counter): New variables.
(process-ambiguities): New Lisp function.
When collecting symbol/jump-tag associations, collect
the lists of jump tags rather than just the first jump tag.
Then process the hash using process-ambiguities to
handle the cases with multiple entries, which are prepared
in the disamb hash. The contents of the disamb hash turn
into HTML material at the end of the page.
An in-line style sheet is now added; it adds vertical space
between the navigation menus so they appear as if they
were separate web pages.