Handle structs in list collection operations.
This eliminates the error which occurs when
a sequence implemented as a struct is nconced
or appended onto. The struct is converted to
a list, and so is the object on the right hand
side. As a result, more generic sequence code
works on structs.

* lib.c (list_collect_nconc, list_collect_append):
Handle tail of type COBJ by converting to a list,
storing that list back in the tail and then
destructively tacking onto that the obj
operand, also turned into a list.