Handle two possible home dir locations on Cygwin.
Under the Cygwin port, if we are running in the Cygwin shell
enviroment, there is a HOME variable which contains a POSIX
path like /home/kaz. This path works fine and we use that.

If we are not running in the Cygwin environment, but running
as an application that uses the Cygwin DLL, then the HOME
variable is still defined, but the path doesn't resolve.

We must detect this, and fall back on the USERPROFILE
environment variable, which holds something like C:\Users\kaz.

* parser.c (get_home_path): New static function,
defined separately for Cygwin and other platforms.
(repl): Use get_home_path to determine home directory.