Fix command spawning on Cygwin-based standalone image.
In the standalone image, we need a /bin/sh interpreter
for the sh and open-command functions to work. The / path
resolves to the installation directory, thanks to some
sysrooting logic in the Cygwin DLL. Thus, we just need
a sh.exe program in the bin directory, side by side
with txr.exe. This sh.exe program can translate the
"sh -c command" invocation to "cmd.exe /c command".

* Makefile (sh.exe): New target, built from cmdwrap.c.

* cmdwrap.c: New file.

* inst.nsi: Install sh.exe.