Don't open text files using "t" mode on Cygwin.
Retracting the recent change to add the "t" mode when
opening non-binary streams on Cygwin, which arranges
for line ending conversion. This restores the Unix-like
treatment of text files on Cygwin, which is expected
of programs.

The Windows native version of TXR will do line ending
conversion thanks to the behavior of text streams in the
Cygnal fork of Cygwin.

* stream.c (format_mode): Only add the "t" option
on Cygwin if compatibility with 144 and 145 is selected.

* txr.1: Updated compatibility notes.