Fix multi-arch build issue for txr.res.
When building in a separate directory, we can't just symlink
the win directory, because the txr.res output file is architecture
dependent and cannot be shared between 32 and 64 bit builds.

* Makefile (%.res): Convert .rc -> .res rule to use the
top_srcdir pattern, so the prerequisite will be found without
the aid of the win/ directory symlink.

* configure: Do not create win -> $source_dir/win
symlink in the build directory.