Fix runaway recursion in lazy struct initialization.
Staci-blowing test case:

   (defstruct foo nil bar)

   (mlet ((f (lnew foo bar (not f.bar))))
     (prinl f.bar))

* struct.c (lazy_struct_init): Do not flip the lazy flag to zero
here. The problem is that it's being done after the
funcall(so->slot[0]), and so the struct is still marked for
lazy initialization while that function is running.
We could detect the circularity (as done in the force function) but that
would create inflexibilities in lazy struct initialization.
(check_init_lazy_struct): Flip the lazy flag to zero in
this function before calling lazy_struct_init.

* txr.1: Document behavior of struct being freely accessible during
lazy initialization.