doc: revise synopsis; document listener invocation.
* txr.1: Revise the synopsis so that it doesn't insinuate
that the query-file or data-file arguments are mandatory.
A formatting issue with the brackets around options is
fixed. Under ARGUMENTS AND OPTIONS document that if
no arguments are present, the interactive listener
is launched. Under INTERACTIVE LISTENER, mention that
not only does the -i option run the listener, but so does
txr invocation with no arguments.