doc: italicize meta-identifers in nroff mode.
* txr.1 (meta, metn, gets): In nroff mode, don't just
add angle brackets around meta symbols, but also typeset
them in italic using \fI...\fP.  This is useful because
the output of man can be colorized when viewed using the
less pager. If these symbols are italicized, then they
get colored differently from surrounding text.
We keep the angle brackets because it helps these words
stand out as "meta" in the absence of color.