awk macro: revise how implicit block works, and name.
* share/txr/stdlib/awk.tl (sys:awk-expander): Check
that :name designates a symbol, and that it isn't nil.
(awk): Move the implicit block to the outermost scope
so it encloses all of the clauses. Default to the
name awk for the block, rather than nil.

* txr.1: Document that the implicit awk block is
called awk by default, and that nil is not allowed
as a block name.