regex: fix broken complement operator.
The form (match-regex "xy" #/~ab/) should return 2 (full
match) because "xy" is in the complement of the set { "ab" }.
It wrongly returns 1.

* regex.c (reg_derivative): Handle the case when
the derivative of the complement's constituent expression
yields nil.  This means that the complemented regex matches
the input.  In this case, the complement must lapse to the .+
regex: match one or more characters.  That is to say, if the
input has at least one more character, there is a match, which
covers all such characters. Otherwise there is no match: the
input matches the complemented regex. In the t case, the
return value is also wrong. If the complemented regex hits
a brick wall (matches nothing, not even the empty string),
the correct complement is "match everything": the .* regex.
Not the match empty string regex!