Bugfix in split-str: empty-match regexes.
* lib.c (split_str_keep): In the regex case, changing
to an infinite loop. The do/while is no longer needed because
the if statement includes a test of the position having
reached the end of the string. This is done before it is
incremented by len, so we avoid wrongly keeping a separator.

* txr.1: Clarified that an empty regex match
behaves like an sep which is an empty string,