Switch regex type from sys:regex to regex.
The sys:regex internal symbol was historically used when
derivative-based regexes had a S-exp representation headed
by that symbol. It had to be a private symbol.
Now the regex symbol is only used as the COBJ class type
for regexes; it makes no sense for it to be in the sys
package. We want user code to be able to refer to this
type using a public symbol.

* lib.c (obj_init): Intern the regex symbol stored
in regex_s in user_package.

* txr.1: Include regex in the type graph in the
Object Type section, and in the type list under
the typeof function.