Bugfix: -Dvar=val not seen in some Lisp code.
TXR Lisp files run from the command line do not see -Dvar=val
bindings, whereas -p expressions do.  The REPL sees the
bindings, but not code loaded from it using (load "file.tl")
because they are lexical.

Let's keep these bindings as local lexicals for -p and -e
forms, but install them as global lexicals for the
other situations.

* parser.c (repl): Get rid of the local repl_env made
from the bindings that are passed in. Instead, before
starting the REPL, loop through the bindings and install
them as global lexicals with reg_varl.

* txr.c (txr_main): Before processing a Lisp file,
install the bindings as global lexicals with reg_varl.