awk macro: loop uses closure to read records.
* share/txr/stdlib/awk.tl (sys:awk-state loop): Refactor
code so that record-processing loop obtains a lambda
which is called to extract a record. If rs and krs did
not change, then the same lambda is obtained from
a variable which caches it. The get-rec-reader local
function thus encapsulates the whole record delimiting
strategy as well as changes to the variables. With this
structure, it will be relatively easy to add support
for Awk's paragraph mode, which is notably missing in
the implementation.