linenoise: bugfix: handle ESC in extended command.
If the user types Ctrl-X and then an arrow key,
we don't want garbage. We let an ESC received
in Ctrl-X extended command mode to be processed
normally and cancel that mode.

* linenoise.c (edit): Bypass the extended command
switch if the character is ESC. Then down below
in the ESC case of the regular command switch,
clear extended mode.