Bugfix: wrong memset size in args_cat_zap.
* args.c (args_copy): Use from->arg and from->fill
in calculating the memcpy size. It doesn't matter
in this function because the to and from are the same;
however, this may be the origin of the copy and paste
error that led to args_cat_zap problem.
(args_copy_zap): Similar change to args_copy: be
consistent in using the from side expressions.
(args_cat_zap): Bugfix: the total size of the
to arguments was used to zap the from side.
This writes zeros past the end of the from
arguments. Fixing by calculating the size in one
place and using the calculated size.