Bugfix: unable to assign to x.y.z place.
This regression was caused by commit
957f80f "Bugfix: issue with expansion of place macros"
on Sep 7, 2016. The commit itself is sound, but exposes a
hidden problem in nearby code.

* share/txr/stdlib/place.tl (sys:pl-expand): The conditions
for terminating the loop and returning the expansion are too
weak, due to the inclusion of an incorrect test.  We must not
bail when the expansion of a compound form is a compound form
with the same symbol. This is because some macros behave that
way, such as, oh: qref!  The expansion of (qref a b c) is
(qref (slot a 'b) 'c): another qref form. To fully expand, we
must keep iterating until the returned form is eq to the input
form.  The original macroexpand (which was replaced by
macroexpand-1 in 957f80f) hid this problem because macroexpand
doesn't use this broken termination test.