Show location of expanded form in exp-time errors.
Old behavior:

  | 1> (sys:expand '(defstruct foo bar))
  | ** defstruct: inheritance base bar does
  |    not name a struct type
  | ** during expansion at
  |    /usr/local/share/txr/stdlib/struct.tl:120
  |    of form (defstruct foo bar)

New behavior:

  | 1> (sys:expand '(defstruct foo bar))
  | ** defstruct: inheritance base bar does
  |    not name a struct type
  | ** during expansion at expr-1:1 of form (defstruct foo
  |                                           bar)
  | ** by macro code located at
  |    /home/kaz/txr/share/txr/stdlib/struct.tl:120

* eval.c (error_trace): Show location of the form being
expanded in the "during expansion" message, rather than,
confusingly, the locaton of the code of its macro.  Then, if
the location of the macro is available, show that in a
separate message whose wording makes it clear that the location
of the expanding macro is being given.