Fix some gc-unsafe mutations found by inspection.
* eval.c (force): When replacing the promise by a
forced value, we must use the set macro.  Only the deref
assignments which store symbols are safe, not the one storing
ret.

* lib.c (alist_nremove, alist_nremove1): We must
use the set macro here instead of assigning through deref.
Even though these assignments preserve the direction of the
list (they just splice out nodes), it's possible that the list
already contains a "wrong-way" reference (old generation to
new) and that the node making this reference is appropriately
marked to be processed properly in the next GC cycle. If we
remove *that* node, we then cause its predecessor to point to
the new generation node and that predecessor could be old
generation.