Fix spectacular bug in number of syntactic places.
Test case: (symacrolet ((f g)) (set [f x] y)) fails to expand.
The root cause is that f is not expanded, because a nil
environment is passed down to with-update-expander macro.
This problem affects not only the dwim place, but nthcdr,
sub and others.

This bug was surprising at first because the Awk macro
establishes the field array f via symbol macro. But
assignments like (set [f 0] ...) worked!

The thing is, they only worked by fluke due to the
way expansions are handled in the Awk macro. That fluke
came to an end in the November 4 commit
157fd76ee20125f409929b95b768b931268b43cf, just before 156
was released.  The changes there don't cause this problem,
but they remove the circumstances which protect the awk
macro from the problem.  So yes, that commit effectively
caused another regression in 156: assignment to awk fields
not working.

* share/txr/stdlib/place.tl (nthcdr, nthlast, ref, sub,
dwim): Do not pass an environment value of nil to
with-update-expander! Pass the value of the special variable
sys:*pl-env* which is there exactly for this purpose:
to give to place expanders like these the original macro
environment where the form occurs, so they can expand
embedded places.